Oops. Left some print statements in that shouldn't have been. As a result, some messages are being printed to the console that shouldn't be and will eventually flood the window (and you can no longer see the menu as a result).
Although you can still use the menu, these statements should be kept out of the production version anyways. The only reason they existed in the first place is because I wrote the Logger class later in the project, and was initially just printing to help debug the filters. I removed most statements but clearly missed a few.