Create SOURCE
Should probably include this in release files, as the GPLv3 license requires making available the source code to any recipients of the program. Users who re-distribute this program might not link to the original source and as a result, may be violating the GPLv3 license. By including a SOURCE file, re-distribution is much less likely to violate GPLv3 as anybody to receives the program will also receive the SOURCE file (so long as those re-distributing don't delete the file).